requirements/general.txt
-r test.txt

Django==3.1.3
python-dotenv==0.15.0
sentry-sdk==1.0.0
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.3
PyYAML==5.3.1
uritemplate==3.0.1
requests==2.25.0
redis==3.5.3
yandex-checkout==1.6.5
qrcode==6.1
Pillow==8.0.1
sorl-thumbnail==12.7.0
